3_Agentic_Rag/requirements.txt
crewai
streamlit
openai
unstructured
langchain
langchain-community
langchain-core
langchain-groq
tools
requests
fastapi
uvicorn
pydantic
python-dotenv
langchain-openai
markitdown
firecrawl-py
pypdf
qdrant-client[fastembed]
langchain-text-splitters